fix(iuserfactory): move IuserFactory interface to the correct directory
- Relocate IuserFactory interface to the appropriate directory for better organization and structure

Refs #20
