+++ backend/src/domain/factories/userFactory.ts
@@ -1,5 +1,5 @@
 import { CreateUserParams } from "@/application/interfaces/domain/entities/user/IauthUser";
-import { IuserFactory } from "@/application/interfaces/domain/entities/factories/IuserFactory";
+import { IuserFactory } from "@/application/interfaces/domain/factories/IuserFactory";
 import { PasswordHash } from "@/application/interfaces/utils/passwordHash";
 import { User } from "../entities/user";
